color: move colour helper functions
Move two colour helper functions closer to their callers.

- color_xterm256_to_24bit()
- modify_color_by_prefix()
